Try to build Ruby project
