+++ rust/optify_ruby/.gitignore
@@ -1,0 +1,1 @@
+*.gem
