+++ rust/optify_ruby/Gemfile
@@ -1,0 +1,6 @@
+source "https://rubygems.org"
+
+gemspec
+
+gem "rake"
+gem "rake-compiler"
