+++ rust/optify_ruby/ext/optify/.cargo/config.toml
@@ -1,0 +1,10 @@
+[build]
+# Without this flag, when linking static libruby, the linker removes symbols
+# (such as `_rb_ext_ractor_safe`) which it thinks are dead code... but they are
+# not, and they need to be included for the `embed` feature to work with static
+# Ruby.
+rustflags = ["-C", "link-dead-code=on"]
+
+# We avoid this on windows due to https://github.com/rust-lang/rust/issues/90056
+#[target.'cfg(not(target_family = "windows"))']
+#rustflags = ["-C", "link-dead-code=on"]
