+++ rust/optify_ruby/ext/optify/Cargo.toml
@@ -16,5 +16,13 @@ exclude = [
     "tests/*",
 ]
 
+[lib]
+crate-type = ["cdylib"]
+
 [dependencies]
 magnus = "0.7.1"
+# optify = { path = "../optify", version = "0.1.0" }
+#rb-sys = { version = "0.9.109", features = [
+#  "global-allocator",
+#] }
+rb-sys = { version = "*", default-features = false, features = ["ruby-static"] }
