+++ rust/optify_ruby/ext/optify/extconf.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+# typed: strict
+
+require "mkmf"
+require "rb_sys/mkmf"
+
+create_rust_makefile("optify/optify")
