+++ rust/optify_ruby/ext/optify/src/lib.rs
@@ -1,0 +1,26 @@
+pub fn add(left: u64, right: u64) -> u64 {
+    left + right
+}
+
+use magnus::{function, Error, Ruby};
+
+fn distance(a: (f64, f64), b: (f64, f64)) -> f64 {
+    ((b.0 - a.0).powi(2) + (b.1 - a.1).powi(2)).sqrt()
+}
+
+#[magnus::init]
+fn init(ruby: &Ruby) -> Result<(), Error> {
+    ruby.define_global_function("distance", function!(distance, 2));
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn it_works() {
+        let result = add(2, 2);
+        assert_eq!(result, 4);
+    }
+}
