+++ rust/optify_ruby/lib/optify.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require_relative "optify/optify"
